Moved styled elements to separate files.
