+++ src/Components/App.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const StyledMain = styled.div`
+    width: 100wv;
+    height: 100hv;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+`;
