+++ src/Components/App.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom';
-import styled from 'styled-components';
+import { StyledMain } from './App.styled';
 import MainArea from './MainArea/MainArea';
 import Navbar from './Navbar/Navbar';
 
@@ -14,11 +13,3 @@ const App = () => {
 };
 
 export default App;
-
-const StyledMain = styled.div`
-    width: 100wv;
-    height: 100hv;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-`;
