+++ src/Components/ErrorPages/404_page.styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const StyledNoMatch = styled.div`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    & > h1 {
+        color: red;
+        font-size: 10rem;
+    }
+`;
