+++ src/Components/Navbar/Navbar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
+import { HeaderLink, StyledHeader } from './Navbard.styled';
 
 const Navbar = () => (
     <StyledHeader>
@@ -17,26 +17,3 @@ const Navbar = () => (
 );
 
 export default Navbar;
-
-const StyledHeader = styled.ul`
-    display: flex;
-    width: 100vw;
-    border-bottom: 2px solid darkolivegreen;
-    background-color: #0f0;
-    margin: 0;
-    padding: 0;
-`;
-
-const HeaderLink = styled.li`
-    list-style-position: outside;
-    list-style-type: none;
-    padding: 1rem;
-    font-size: 2rem;
-    font-weight: bold;
-    color: black;
-    user-select: none;
-
-    &:hover {
-        text-decoration: underline;
-    }
-`;
